Separates bash script creation and network folder creation...
- switches to mostly using relative paths in the scripts
- $CONSENSUS-genesis.json changed to just genesis.json everywhere
- Adds some extra log lines to start script
- Mostly leaves the bash script creation logic the same for now, but this needs to be cleaned up
