+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "quorum-creator",
-  "version": "0.1.1",
+  "version": "0.1.2",
   "description": "Tool for walking through the setup of Quorum networks.",
   "main": "src/index.js",
   "bin": {
