+++ src/generators/bashHelper.test.js
@@ -1,64 +1,112 @@
-import { createDirectory } from './networkCreator'
-import { createReplica7NodesConfig } from '../model/NetworkConfig'
-import { buildBashScript, buildBash, waitForTesseraNodesCommand } from './bashHelper'
-import { copyFile, cwd, libRootDir, writeFile } from '../utils/fileUtils'
-import { execute } from '../utils/execUtils'
+import { buildBashScript } from './bashHelper'
+import { createCustomConfig, createQuickstartConfig, createReplica7NodesConfig } from '../model/NetworkConfig'
+import { cwd, libRootDir } from '../utils/fileUtils'
 import { TEST_CWD, TEST_LIB_ROOT_DIR } from '../utils/testHelper'
-import { pathToCakeshop, pathToQuorumBinary, pathToTesseraJar } from './binaryHelper'
+import { generateAccounts } from './consensusHelper'
 
-jest.mock('../generators/networkCreator')
 jest.mock('../utils/fileUtils')
-jest.mock('../utils/execUtils')
-jest.mock('../generators/binaryHelper')
+jest.mock('../generators/consensusHelper')
 cwd.mockReturnValue(TEST_CWD)
 libRootDir.mockReturnValue(TEST_LIB_ROOT_DIR)
-pathToQuorumBinary.mockReturnValue('geth')
-pathToTesseraJar.mockReturnValue('tessera')
-pathToCakeshop.mockReturnValue('cakeshop')
+generateAccounts.mockReturnValue("accounts")
 
-describe('generates bash script details', () => {
-  it('creates bash script given config details', () => {
-    let config = createReplica7NodesConfig({
-      numberNodes: '5',
-      consensus: 'raft',
-      quorumVersion: '2.4.0',
-      transactionManager: '0.10.2',
-      deployment: 'bash',
-      cakeshop: false
-    })
-    const expected = {
-      startScript: `BIN_GETH=geth\nBIN_TESSERA=tessera\n\nstartTessera\n${waitForTesseraNodesCommand(config)}\nstartGeth\n`,
-      initCommands: ['1', '2', '3', '4', '5'],
-      networkPath: 'testPath'
-    }
-    createDirectory.mockReturnValueOnce({tesseraStart:  "startTessera",
-        gethStart: "startGeth",
-        initStart: ['1', '2', '3', '4', '5'],
-        netPath: "testPath",
-      })
-    expect(buildBashScript(config)).toEqual(expected)
-  })
+test('creates quickstart config', () => {
+  const config = createQuickstartConfig()
+  const bash = buildBashScript(config).startScript
+  expect(bash).toMatchSnapshot()
 })
 
-describe('builds bash directory', () => {
-  it('given bash details builds files to run bash', async () => {
-    let config = createReplica7NodesConfig({
-      numberNodes: '5',
-      consensus: 'raft',
-      quorumVersion: '2.4.0',
-      transactionManager: '0.10.2',
-      deployment: 'bash',
-      cakeshop: false
-    })
-    createDirectory.mockReturnValueOnce({tesseraStart:  "startTessera",
-        gethStart: "startGeth",
-        initStart: ['1', '2', '3', '4', '5'],
-        netPath: "test",
-      })
-    await buildBash(config)
-
-    expect(writeFile).toBeCalledWith('test/start.sh', expect.any(String), true)
-    expect(copyFile).toBeCalledTimes(5)
-    expect(execute).toBeCalledTimes(5)
+test('creates 3nodes raft bash tessera', () => {
+  const config = createReplica7NodesConfig({
+    numberNodes: '3',
+    consensus: 'raft',
+    quorumVersion: '2.4.0',
+    transactionManager: '0.10.2',
+    deployment: 'bash',
+    cakeshop: false
   })
+  const bash = buildBashScript(config).startScript
+  expect(bash).toMatchSnapshot()
+})
+
+test('creates 3nodes raft bash tessera cakeshop', () => {
+  const config = createReplica7NodesConfig({
+    numberNodes: '3',
+    consensus: 'raft',
+    quorumVersion: '2.4.0',
+    transactionManager: '0.10.2',
+    deployment: 'bash',
+    cakeshop: true
+  })
+  const bash = buildBashScript(config).startScript
+  expect(bash).toMatchSnapshot()
+})
+
+test('creates 3nodes raft bash tessera custom', () => {
+  const config = createCustomConfig({
+    numberNodes: '3',
+    consensus: 'raft',
+    quorumVersion: '2.4.0',
+    transactionManager: '0.10.2',
+    deployment: 'bash',
+    cakeshop: false,
+    generateKeys: false,
+    networkId: 10,
+    genesisLocation: 'none',
+    customizePorts: false,
+    nodes: [],
+    dockerCustom: undefined
+  })
+  const bash = buildBashScript(config).startScript
+  expect(bash).toMatchSnapshot()
+})
+
+test('creates 2nodes istanbul bash tessera cakeshop custom ports', () => {
+  let nodes = [
+    {
+      quorum: {
+        ip: '127.0.0.1',
+        devP2pPort: '55001',
+        rpcPort: '56000',
+        wsPort: '57001',
+        raftPort: '80501',
+      },
+      tm: {
+        ip: '127.0.0.1',
+        thirdPartyPort: '5081',
+        p2pPort: '5001',
+        enclavePort: '5181',
+      }
+    },
+    {
+      quorum: {
+        ip: '127.0.0.1',
+        devP2pPort: '55001',
+        rpcPort: '56001',
+        wsPort: '56001',
+        raftPort: '80502',
+      },
+      tm: {
+        ip: '127.0.0.1',
+        thirdPartyPort: '5082',
+        p2pPort: '5002',
+        enclavePort: '5182',
+      }
+    }]
+  const config = createCustomConfig({
+    numberNodes: '2',
+    consensus: 'istanbul',
+    quorumVersion: '2.4.0',
+    transactionManager: '0.10.2',
+    deployment: 'bash',
+    cakeshop: true,
+    generateKeys: false,
+    networkId: 10,
+    genesisLocation: 'none',
+    customizePorts: true,
+    nodes: nodes,
+    dockerCustom: undefined
+  })
+  const bash = buildBashScript(config).startScript
+  expect(bash).toMatchSnapshot()
 })
