+++ src/questions/index.js
@@ -24,6 +24,7 @@ import {
 import inquirer from 'inquirer'
 import { cwd, readFileToString } from '../utils/fileUtils'
 import { join } from "path"
+import { createDirectory } from '../generators/networkCreator'
 
 export async function quickstart() {
   const config = createQuickstartConfig()
@@ -79,6 +80,7 @@ export async function customize () {
 
 async function buildNetwork(config, deployment) {
   console.log('')
+  createDirectory(config)
   if (deployment === 'bash') {
     await buildBash(config)
   } else if (deployment === 'docker-compose') {
